requirements.txt
flask
numpy
pymongo
dnspython
python-dotenv
